refactor time!
Created a dedicated Changelog struct which now has all the crazy parts
in it. This cleans up the main.rs file where we just call methods on the
changelog, and this also cleans up the ast file because now all the
domain logic of an actual Changelog is moved. This is now purely an AST
for a MD file.